+++ tests/test_datasets.py
@@ -1,6 +1,6 @@
 import pytest
 
-from timemachine.datasets import fetch_freesolv
+from timemachine.testsystems import fetch_freesolv
 
 pytestmark = [pytest.mark.nocuda]
