--- timemachine/datasets/__init__.py
@@ -1,3 +1,0 @@
-from timemachine.datasets.utils import fetch_freesolv
-
-__all__ = ["fetch_freesolv"]
