+++ timemachine/testsystems/__init__.py
@@ -1,3 +1,4 @@
 from . import ligands, relative
+from .utils import fetch_freesolv
 
-__all__ = ["ligands", "relative"]
+__all__ = ["fetch_freesolv", "ligands", "relative"]
