+++ timemachine/testsystems/freesolv/readme.md
@@ -1,0 +1,4 @@
+## `freesolv`
+`freesolv.sdf` contains a sanitized version of v0.52 the FreeSolv database of hydration free energies curated by David Mobley, Peter Guthrie, and colleagues here: https://github.com/MobleyLab/FreeSolv/releases/tag/v0.52
+
+(Note: `freesolv.sdf` reflects a fix described in https://github.com/MobleyLab/FreeSolv/issues/48 )
